+++ src/deploy/AWSDeployer.js
@@ -54,6 +54,7 @@ class AWSDeployer extends BaseDeployer {
     super(builder);
 
     Object.assign(this, {
+      name: 'AWS',
       _region: '',
       _role: '',
       _functionARN: '',
@@ -87,6 +88,12 @@ class AWSDeployer extends BaseDeployer {
 
   get host() {
     return `${this._apiId}.execute-api.${this._region}.amazonaws.com`;
+  }
+
+  validate() {
+    if (!this._role || !this._region) {
+      throw Error('AWS target needs --aws-region and --aws-role');
+    }
   }
 
   async init() {
@@ -264,7 +271,7 @@ class AWSDeployer extends BaseDeployer {
 
     const { ApiId, ApiEndpoint } = res;
     this._apiId = ApiId;
-    this._functionURL = ApiEndpoint;
+    this._functionURL = `${ApiEndpoint}/${this._builder.actionName.replace('@', '_')}`;
 
     // check for stage
     res = await this._api.send(new GetStagesCommand({
@@ -319,8 +326,6 @@ class AWSDeployer extends BaseDeployer {
       StatementId: crypto.randomBytes(16).toString('hex'),
     }));
 
-    this._functionURL += `/${this._builder.actionName.replace('@', '_')}`;
-
     if (this._builder.showHints) {
       const opts = '';
       this.log.info('\nYou can verify the action with:');
@@ -336,6 +341,10 @@ class AWSDeployer extends BaseDeployer {
       url: this._functionURL,
       idHeader: 'apigw-requestid',
     });
+  }
+
+  get fullFunctionName() {
+    return this._functionURL;
   }
 
   async updatePackage() {